fix filters styles
