+++ client/src/containers/projects/list.tsx
@@ -10,6 +10,8 @@ import ProjectItem from '@/containers/projects/item';
 
 import { Input } from '@/components/ui/input';
 
+import FiltersSelected from '../filters/selected';
+
 export default function ProjectsList() {
   const [searchValue, setSearchValue] = useState('');
 
@@ -32,16 +34,16 @@ export default function ProjectsList() {
 
   return (
     <div className="no-scrollbar flex max-h-[75vh] flex-col space-y-2 overflow-y-auto">
-      <div className="relative flex justify-between">
+      <div className="relative mx-1 mt-1 flex justify-between space-x-2">
         <Input
           placeholder="Search project by name"
           className="border-none bg-gray-100 pl-11 placeholder:text-gray-500"
           onChange={(e) => setSearchValue(e.target.value)}
         />
         <Search size={24} className="absolute left-2 top-2" />
-
         <Filters />
       </div>
+      <FiltersSelected />
       {data && data.map((project) => <ProjectItem key={project?.id} data={project} />)}
     </div>
   );
